Workable Prototype + Cardinal Change
Major format change to how the World is handled. Added Cardinal System which is constantly checking Player, Creature, and NPC loaded Areas, and updates World DATA. This version now has UniqueID for each NPC, Creature, and Item, this makes tracking and handling a lot easier. This form is currently in a Workable Prototype stage. You are able to join, and battle a Goblin with other Players. This is the most barebones state of the Game, and is now in a state to actually create new game content.